Fix for special character formatting bug
While playing the game, I discovered that formatting for answers that include special characters was inconsistent and almost always had problems with hyphenated answers and apostrophes. 

I wrote a few tests that validated/replicated the bug I was seeing. I discovered that if all characters except for numbers and letters are replaced with an empty string, instead of listing the characters to be removed, the tests pass and the problem is resolved.